+++ Wiki/Azure/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2021.11.16]
+
+## Updated
+
+- Changed and added functionality to the `Create-CosmosDbAccount.ps1` script. Functionality has been added for: `Table`, `Gremlin` and `MongoDb`.
+- Added the ability to pick a set of diagnostic settings for every resource.
+- Added the ability to disable diagnostic settings for every resource.
+
 ## [2021.11.01]
 
 ## Breaking Changes